tests for valid api version in request-url
